pyproject.toml
[project]
name = "lb-backend"
version = "0.1.0"
description = "LocationsBook backend API"
requires-python = ">=3.13,<3.14"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "sqlalchemy[asyncio]",
    "asyncpg",
    "alembic",
    "pydantic",
    "pydantic-settings",
    "pyjwt[crypto]",
    "bcrypt",
    "email-validator",
    "python-multipart",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
    "mypy",
    "types-pyjwt",
]

[tool.ruff]
target-version = "py313"
line-length = 88

[tool.ruff.lint]
select = ["E", "W", "F", "I", "N", "UP", "B", "SIM", "RUF"]
per-file-ignores = {"app/api/**" = ["B008"]}

[tool.mypy]
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "asyncpg.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "bcrypt.*"
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
